Fine tuning for schedule sorting

+++ Lynx/Rest API/Schedule.swift
@@ -14,8 +14,8 @@ struct ScheduleSearchOptions {
     var searchFullCourses: Bool = true
     var gapOrder: Order = .asc
     var dayOrder: Order = .desc
-    var earliest: Double = 700
-    var latest: Double = 2300
+    var earliest: Int = 700
+    var latest: Int = 2300
 }
 
 struct Schedule: Codable, Equatable {
